[ALPS02968842] cmdq: Enable 64bit handle
[Detail]
Change handle and PA to 64 bit.

MTK-Commit-Id: 619602cebf77dffefa2da7c9d4209c47d66c3863

CR-Id: ALPS02968842
Feature: CMDQ

+++ drivers/misc/mediatek/cmdq/v3/inc/cmdq_sec_dr/cmdq_sec_iwc_common.h
@@ -61,7 +61,7 @@ struct iwcCmdqAddrMetadata_t {
 	 */
 
 	uint32_t type;			/* [IN] addr handle type*/
-	uint32_t baseHandle;	/* [IN]_h, secure address handle */
+	uint64_t baseHandle;	/* [IN]_h, secure address handle */
 	uint32_t blockOffset;	/* [IN]_b, block offset from handle(PA) to current block(plane) */
 	uint32_t offset;		/* [IN]_b, buffser offset to secure handle */
 	uint32_t size;			/* buffer size */
